requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.3
Pillow==10.3.0
pytesseract==0.3.10
requests==2.32.3
python-dotenv==1.0.1
